[댓글 기능] 버디 미션 제출합니다.  (#118)
* docs: 댓글 기능 명세

* feat: Interceptor구현

* fix: Controller,Service HttpSession 중복 제거

* fix: Interceptor 기능 수정

* feat: Article과 User의 관계 매핑 및 Interceptor 적용

* feat: comment 작성 구현

* feat: Comment 삭제 기능 구현

* feat: Comment 수정 기능 구현

* feat: ArticleControllerTest 테스트 케이스 추가 및 data.sql 추가

* feat: data.sql 추가

* feat: ArticleTest 추가

* refactor: ArticleController의 댓글 로직 ConmmentController로 분리 및 테스트 이동

* feat: comment 기능 추가 및 테스트 추가

* feat: Article,Comment Class의 contents필드에 @Lob어노테이션 추가

* refactor: ArticleController, ArticleService 리팩토링 및 테스트 추가

* feat: Date Class 생성 및 객체 상속

* refactor: 테스트에 사용되는 공용 필드 static으로 변경 및 리팩토링

* fix: LoginService의 login 메서드 filter로 변경 및 테스트 코드 추가

* refactor: 테스트 코드 전체 리팩토링

* feat: DataJpaTest 추가

* refactor: ExceptionHandler 메서드 공통 핸들러 클래스로 분리

+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -3,11 +3,13 @@ package techcourse.myblog.service;
 import org.springframework.stereotype.Service;
 import techcourse.myblog.article.Article;
 import techcourse.myblog.article.ArticleRepository;
+import techcourse.myblog.comment.Comment;
 import techcourse.myblog.exception.NotFoundObjectException;
 import techcourse.myblog.service.dto.ArticleDto;
+import techcourse.myblog.user.User;
 
-import javax.servlet.http.HttpSession;
 import javax.transaction.Transactional;
+import java.util.List;
 
 @Service
 public class ArticleService {
@@ -17,11 +19,9 @@ public class ArticleService {
         this.articleRepository = articleRepository;
     }
 
-    public Article createArticle(ArticleDto articleDto, HttpSession httpSession) {
-        if(httpSession.getAttribute("user")== null){
-            throw new NotFoundObjectException();
-        }
-        Article article = articleDto.toEntity();
+    public Article createArticle(ArticleDto articleDto, User author) {
+        Article article = articleDto.toEntity(author);
+
         return articleRepository.save(article);
     }
 
@@ -38,7 +38,23 @@ public class ArticleService {
         return article;
     }
 
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(Long articleId, User user) {
+        Article article = findArticle(articleId);
+        article.checkCorrespondingAuthor(user);
         articleRepository.deleteById(articleId);
+    }
+
+    public List<Comment> findAllComments(Article article) {
+        return article.getComments();
+    }
+
+    @Transactional
+    public void addComment(Long articleId, Comment comment) {
+        Article article = articleRepository.findById(articleId).orElseThrow(NotFoundObjectException::new);
+        article.addComment(comment);
+    }
+
+    public void checkAvailableUpdateUser(Article article, User user) {
+        article.checkCorrespondingAuthor(user);
     }
 }
